fixing more bugs with attacking

+++ Fighters/Stats.java
@@ -3,7 +3,7 @@ package Fighters;
 import java.util.*;
 
 public class Stats {
-  private EnumMap<Attribute, Double> attributes;
+  private EnumMap<Attribute, Integer> attributes;
   private Set<Attribute> favoredAttributes;
 
   public Stats(Set<Attribute> favoredAttributes) {
@@ -17,7 +17,7 @@ public class Stats {
    * @param attr  selected attribute
    * @param value new value for it
    */
-  public void set(Attribute attr, double value) {
+  public void set(Attribute attr, int value) {
     attributes.put(attr, value);
   }
 
@@ -25,10 +25,10 @@ public class Stats {
    * get value of an attribute or 0.0 if it doesn't exists
    * 
    * @param attr selected attribute used as key
-   * @return its value or 0.0
+   * @return its value or 0
    */
-  public double get(Attribute attr) {
-    return attributes.getOrDefault(attr, 0.0);
+  public int get(Attribute attr) {
+    return attributes.getOrDefault(attr, 0);
   }
 
   /**
@@ -39,7 +39,7 @@ public class Stats {
   public void levelUp() {
     double baseIncrease = 0.05;
     for (Attribute attr : attributes.keySet()) {
-      double value = attributes.get(attr);
+      int value = attributes.get(attr);
       value *= attr == Attribute.MANA ? 1.1 : (1 + baseIncrease); // mana points scale by 1.1, everything else by 1.05
       if (favoredAttributes.contains(attr)) {
         value *= (1 + baseIncrease);
